Make `currents` only accessible via a helper method
`currents`, today, is a bit of a foot-gun since it is 0-index based, and
the channel numbering on devices is 1-index based.  It is very easy to
end up with off-by-one errors when using this library, unless you are
careful and have good test coverage.

This change puts access to the current of each channel behind a helper
method on a `Packet` that mitigates this.