Enable callers to properly close protocol transports
`asyncio` wants transports to be closed. However, when the transports are being created by an asyncio `Server` and passed to new `PacketProtocol` or `BidirectionalProtocol` instances, doing that requires the user of this library to do a lot of obnoxious plumbing of their own.

This PR makes it a little bit easier for them in two ways:
- `PacketProtocol` now sends a message to its associated queue when connections are made and lost, which allows the caller to keep track of active protocol instances
- `PacketProtocol` exposes a `close` method that can be used to close an active protocol

This is a breaking change, since previously the queue contained only `Packet` objects, and now it contains `PacketProtocolMessage`s instead.