Allow customizing the packet clear time interval used by the library
This allows passing a new argument to a `BidirectionalProtocol` that
lets the user configure the time the protocol will wait before
attempting to send a command.

Should help work around the problem discussed in #84

This introduces a breaking change that removes the constant
`PACKET_DELAY_CLEAR_TIME` from `siobrultech_protocols.gem.protocol`.  A
new constant, `PACKET_DELAY_CLEAR_TIME_DEFAULT` is added to
`siobrultech_protocols.gem.const` that is the default value we use if a
value is not provided to the `BidirectionalProtocol`.