Add ECM-1240 and (untested) ECM-1220 support
I have an ECM-1240 (courtesy of Brultech) so I was able to generate
some packet format 3 packets and test this using those.

Unfortunately, proper setup of an ECM-1240 with the CTs that I have
requires a resistor (which I don't have) to be installed in parallel.
So, while I'm reasonably certain this parsing code is correct, it was
producing current and power numbers that were orders of magnitude off
for my test setup.

I don't have ECM-1220 packets to test with, but the ECM-1240 packet spec
included documentation for those as a subset of the ECM-1240 packet, so I
went ahead and defined that packet type too.