Remove PacketProtocol.close()
I originally put this in with the idea that it was difficult for callers to keep track of transports to close them (#51). That's not really accurate, though. For `asyncio`, `loop.create_connection()` returns the transport, and the server returned by `loop.create_server()` has a `sockets` member that keeps track of all open sockets. This code snippet from `greeneye_monitor` is an example of how to close them upon shutdown:

```
            # Close existing connections
            for socket in self._server.sockets:
                socket.shutdown(socket.SHUT_RDWR)
                socket.close()
```