Allow packets to interleave with API responses
I'm working towards ECM-1240 API support. One difference between the GEM and ECM APIs is that the ECM API does not have a way to pause packet sends. As a result, we need to be resilient to a packet arriving while an API call is in progress.

Previously, `BidirectionalProtocol` would intercept all incoming data once it had reached the `SENT_API_REQUEST` state (which is after the packet pause request and the subsequent delay). This PR changes it to only intercept data that `PacketProtocol` does not recognize as a packet.

The code assumes that the packets are sent atomically with respect to API responses -- that is, while a packet might interrupt an API response, no API response can interrupt a packet. That's not documented one way or another, but it seems to be true in practice. (I've actually tried *not* sending the packet pause request to the GEM and this code works fine.)