Return API call results as soon as they arrive (#123)
Our current implementation of API calls just asks the caller to wait a few seconds and then read the response. However, GEM and ECM-1240 API responses have either a predictable length or a predictable terminator, so we can do better.

This PR changes how `BidirectionalProtocol` works. Instead of having separate methods to send a request and receive a reply, it now has one `invoke_api` method which takes an `ApiCall`, the parameters for the call, and a `Future` into which the result will be placed. Callers now `await` that `Future` with a timeout, and typically will receive a response almost immediately.

This is a breaking change for callers who use the protocol directly. Callers who use the higher-level APIs (`api.call_api` or the individual methods) just get faster respones with no code change needed.

This is also a breaking change for callers that define their own `ApiCall`s; the parsers must now handle the case of not having enough data (returning `None` in that case).