Enable packet delay to be skipped (#125)
This PR gives the option of disabling the packet delay, while still leaving it on by default. In my own testing locally, I've seen that we can get away without sending the packet delay request, and that doing so makes GEM API calls almost instantaneous. I'd like to be able to try that out at broader scale, without requiring a new release if it turns out to not work in some cases.