ECM API support
The ECM API is rather different from GEM. Instead of a single round-trip per call, each API call involves multiple round-trips. For example, in order to set the packet send interval:

1. PC sends FC (hex)
2. ECM-1240 responds with FC (hex)
3. PC sends three ASCII bytes “S” “E” “T” (must be upper case characters)
4. ECM-1240 responds with FC (hex)
5. PC sends three ASCII bytes “I” “V” “2” (must be upper case characters)
6. ECM-1240 responds with FC (hex)
7. PC sends a single byte chr$(interval value) sends the ascii character representing the interval.
8. ECM-1240 responds with FC (hex)

So that's 4 round-trips for one API call.

This PR updates `ApiCall` to allow separate formatters and parsers to be supplied for GEM and ECM, adds the necessary behaviors in `BidirectionalProtocol`, and supplies implementations for serial number and packet send frequency (the two GEM APIs we support that have ECM equivalents).